LPD-40735 - Refactor. Rename
Extract interface to shared types
