+++ modules/apps/frontend-data-set/frontend-data-set-admin-web/src/main/resources/META-INF/resources/js/utils/types.ts
@@ -45,6 +45,35 @@ export interface ITable extends IBaseVisualizationMode<'table'> {}
 
 export type TVisualizationMode = ICards | IList | ITable;
 
+export interface IDataSet {
+	actions: {
+		delete: {
+			href: string;
+			method: string;
+		};
+		update: {
+			href: string;
+			method: string;
+		};
+	};
+	additionalAPIURLParameters?: string;
+	creationActionsOrder?: string;
+	defaultItemsPerPage: number;
+	defaultVisualizationMode?: string;
+	description?: string;
+	externalReferenceCode: string;
+	filtersOrder?: string;
+	id: string;
+	itemActionsOrder?: string;
+	label: string;
+	listOfItemsPerPage: string;
+	restApplication: string;
+	restEndpoint: string;
+	restSchema: string;
+	sortsOrder?: string;
+	tableSectionsOrder?: string;
+}
+
 export interface IField {
 	children?: Array<IField>;
 	format?: EFieldFormat;
